# infoschema: fix issue of information schema cache miss cause by schema version gap (#53445)