--- a/util/logutil/BUILD.bazel
+++ b/util/logutil/BUILD.bazel
@@ -16,6 +16,7 @@ go_library(
         "@com_github_opentracing_opentracing_go//log",
         "@com_github_pingcap_errors//:errors",
         "@com_github_pingcap_log//:log",
+        "@com_github_tikv_client_go_v2//tikv",
         "@org_uber_go_zap//:zap",
         "@org_uber_go_zap//buffer",
         "@org_uber_go_zap//zapcore",