Cherry-pick #53445  #47810 

<!--

Thank you for contributing to TiDB!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?
<!--

Please create an issue first to describe the problem.

There MUST be one line starting with "Issue Number:  " and
linking the relevant issues via the "close" or "ref".

For more info, check https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/contribute-code.html#referring-to-an-issue.

-->

Issue Number: close #53428

Problem Summary: fix issue of information schema cache miss cause by schema version gap

### What changed and how does it work?

#### before this PR

When DDL has empty schema version, will cause cache miss, then stale-read query need to load snapshot schema from TiKV.

<img width="1454" alt="image" src="https://github.com/pingcap/tidb/assets/26020263/32411733-9e38-4bf5-9b9e-b0287af18f31">


```sql
[2024/05/23 06:15:10.546 +00:00] [INFO] [domain.go:287] ["diff load InfoSchema success"] [currentSchemaVersion=6760] [neededSchemaVersion=6762] ["start time"=2.583049ms] [gotSchemaVersion=6762] [phyTblIDs="[]"] [actionTypes="[]"] [diffTypes="[\"drop table\"]"]
[2024/05/23 06:15:17.869 +00:00] [INFO] [domain.go:287] ["diff load InfoSchema success"] [currentSchemaVersion=6763] [neededSchemaVersion=6765] ["start time"=2.040666ms] [gotSchemaVersion=6765] [phyTblIDs="[10496]"] [actionTypes="[4]"] [diffTypes="[\"drop table\"]"]
[2024/05/23 06:15:18.051 +00:00] [INFO] [domain.go:287] ["diff load InfoSchema success"] [currentSchemaVersion=6765] [neededSchemaVersion=6767] ["start time"=4.321375ms] [gotSchemaVersion=6767] [phyTblIDs="[]"] [actionTypes="[]"] [diffTypes="[\"drop table\"]"]
[2024/05/23 06:15:18.245 +00:00] [INFO] [domain.go:287] ["diff load InfoSchema success"] [currentSchemaVersion=6767] [neededSchemaVersion=6769] ["start time"=3.01603ms] [gotSchemaVersion=6769] [phyTblIDs="[]"] [actionTypes="[]"] [diffTypes="[\"drop table\"]"]
```

### This PR

When DDL has empty schema version, this PR won't have schema cache miss, then won't cause many load snapshot schema operation.

<img width="1445" alt="image" src="https://github.com/pingcap/tidb/assets/26020263/7d34a3c9-aa04-453b-87c5-51d369b04b43">

```log
[2024/05/23 06:23:45.209 +00:00] [INFO] [domain.go:287] ["diff load InfoSchema success"] [currentSchemaVersion=6777] [neededSchemaVersion=6779] ["start time"=5.409896ms] [gotSchemaVersion=6779] [phyTblIDs="[10502]"] [actionTypes="[4]"] [diffTypes="[\"drop table\"]"]
[2024/05/23 06:23:45.458 +00:00] [INFO] [domain.go:287] ["diff load InfoSchema success"] [currentSchemaVersion=6779] [neededSchemaVersion=6781] ["start time"=2.711853ms] [gotSchemaVersion=6781] [phyTblIDs="[]"] [actionTypes="[]"] [diffTypes="[\"drop table\"]"]
[2024/05/23 06:23:45.650 +00:00] [INFO] [domain.go:287] ["diff load InfoSchema success"] [currentSchemaVersion=6781] [neededSchemaVersion=6783] ["start time"=1.910751ms] [gotSchemaVersion=6783] [phyTblIDs="[]"] [actionTypes="[]"] [diffTypes="[\"drop table\"]"]
[2024/05/23 06:23:49.173 +00:00] [INFO] [domain.go:287] ["diff load InfoSchema success"] [currentSchemaVersion=6784] [neededSchemaVersion=6786] ["start time"=2.511772ms] [gotSchemaVersion=6786] [phyTblIDs="[10505]"] [actionTypes="[4]"] [diffTypes="[\"drop table\"]"]
[2024/05/23 06:23:49.372 +00:00] [INFO] [domain.go:287] ["diff load InfoSchema success"] [currentSchemaVersion=6786] [neededSchemaVersion=6788] ["start time"=2.768253ms] [gotSchemaVersion=6788] [phyTblIDs="[]"] [actionTypes="[]"] [diffTypes="[\"drop table\"]"]
[2024/05/23 06:23:49.590 +00:00] [INFO] [domain.go:287] ["diff load InfoSchema success"] [currentSchemaVersion=6788] [neededSchemaVersion=6790] ["start time"=4.466332ms] [gotSchemaVersion=6790] [phyTblIDs="[]"] [actionTypes="[]"] [diffTypes="[\"drop table\"]"]
[2024/05/23 06:23:53.670 +00:00] [INFO] [domain.go:287] ["diff load InfoSchema success"] [currentSchemaVersion=6791] [neededSchemaVersion=6793] ["start time"=799.389µs] [gotSchemaVersion=6793] [phyTblIDs="[10508]"] [actionTypes="[4]"] [diffTypes="[\"drop table\"]"]
[2024/05/23 06:23:53.872 +00:00] [INFO] [domain.go:287] ["diff load InfoSchema success"] [currentSchemaVersion=6793] [neededSchemaVersion=6795] ["start time"=1.025364ms] [gotSchemaVersion=6795] [phyTblIDs="[]"] [actionTypes="[]"] [diffTypes="[\"drop table\"]"]
[2024/05/23 06:23:54.090 +00:00] [INFO] [domain.go:287] ["diff load InfoSchema success"] [currentSchemaVersion=6795] [neededSchemaVersion=6797] ["start time"=3.549838ms] [gotSchemaVersion=6797] [phyTblIDs="[]"] [actionTypes="[]"] [diffTypes="[\"drop table\"]"]
```


### Check List

Tests <!-- At least one of them must be included. -->

- [x] Unit test
- [ ] Integration test
- [x] Manual test (add detailed scripts or steps below)
- [ ] No need to test
  > - [ ] I checked and no code files have been changed.
  > <!-- Or your custom  "No need to test" reasons -->

Side effects

- [ ] Performance regression: Consumes more CPU
- [ ] Performance regression: Consumes more Memory
- [ ] Breaking backward compatibility

Documentation

- [ ] Affects user behaviors
- [ ] Contains syntax changes
- [ ] Contains variable changes
- [ ] Contains experimental features
- [ ] Changes MySQL compatibility

### Release note

<!-- compatibility change, improvement, bugfix, and new feature need a release note -->

Please refer to [Release Notes Language Style Guide](https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/release-notes-style-guide.html) to write a quality release note.

```release-note
None
```